Implemented MyCart and Shopping

+++ Flask-Backend/application/api/shopping.py
@@ -8,14 +8,11 @@ from flask_restful import Resource
 from flask_restful import fields, marshal
 from flask_restful import reqparse
 from application.data.database import db
-from flask_login import  login_required
-from flask_security import roles_required
-
-mycart_fields = {
-    'cart_id': fields.Integer
-}
+from flask_login import  login_required, current_user
+from flask_security import roles_required, auth_required
 
 incart_product_fields = {
+    'p_id': fields.Integer(attribute='Products.p_id'),
     'p_name': fields.String(attribute='Products.p_name'),
     'p_description' : fields.String(attribute='Products.p_description'),
     'p_qty' : fields.Integer(attribute='Products.p_qty'),
@@ -32,7 +29,7 @@ class MyCartCRUD(Resource):
         self.parser.add_argument('bought_qty', type=int)
 
     @roles_required('customer')
-    @login_required
+    @auth_required('token')
     def get(self, user_id):
         mycart_data = MyCart.query.filter_by(user_id=user_id).first()
         products = mycart_data.hasItems
@@ -43,14 +40,14 @@ class MyCartCRUD(Resource):
                     .filter(ProductCart.cart_id==mycart_data.cart_id, Products.p_id==p.p_id).first()
             incart_product_data.append(new_p)
         result = {
-            'mycart': marshal(mycart_data, mycart_fields),
+            'cart_id': mycart_data.cart_id,
             'products': marshal(incart_product_data, incart_product_fields)
         }
         return result, 200
     
     @roles_required('customer')
-    @login_required
-    def post(self, user_id, p_id):
+    @auth_required('token')
+    def post(self, user_id, p_id):      # add to cart
         args = self.parser.parse_args()
         mycart = MyCart.query.filter_by(user_id=user_id).first()
         prodExist = ProductCart.query.filter_by(p_id=p_id, cart_id=mycart.cart_id).first()
@@ -63,19 +60,19 @@ class MyCartCRUD(Resource):
         return {'message':'Product added to Cart'}, 200
     
     @roles_required('customer')
-    @login_required
-    def put(self, user_id, p_id):
+    @auth_required('token')
+    def put(self, p_id):
         args = self.parser.parse_args()
-        mycart = MyCart.query.filter_by(user_id=user_id).first()
+        mycart = MyCart.query.filter_by(user_id=current_user.id).first()
         product = ProductCart.query.filter_by(p_id=p_id, cart_id=mycart.cart_id).first()
         product.bought_qty = args['bought_qty']
         db.session.commit()
         return 200
 
     @roles_required('customer')
-    @login_required
-    def delete(self, user_id, p_id):
-        mycart = MyCart.query.filter_by(user_id=user_id).first()
+    @auth_required('token')
+    def delete(self, p_id):
+        mycart = MyCart.query.filter_by(user_id=current_user.id).first()
         product = ProductCart.query.filter_by(p_id=p_id, cart_id=mycart.cart_id).first()
         db.session.delete(product)
         db.session.commit()
@@ -90,7 +87,7 @@ transaction_fields = {
 
 class TransactionConfirm(Resource):
     @roles_required('customer')
-    @login_required
+    @auth_required('token')
     def get(self, user_id):
         trans = Transaction.query.filter_by(user_id=user_id).all()
         result = []
@@ -114,7 +111,7 @@ obj.set_count()
 
 class PlaceOrder(Resource):    
     @roles_required('customer')
-    @login_required
+    @auth_required('token')
     def post(self, user_id):
         mycart = MyCart.query.filter_by(user_id=user_id).first()
         products = mycart.hasItems
@@ -124,28 +121,33 @@ class PlaceOrder(Resource):
         total_price = 0
         for p in products:
             pc = ProductCart.query.filter_by(cart_id=mycart.cart_id, p_id=p.p_id).first()
-            # applying category offer to products individually
-            discount, is_discount =0, False
-            offer = Category.query.get(p.c_id).offers
-            if offer is not None:                 
-                discount-=offer.apply_discount(p.price)
-                is_discount=True
-            total_price+=(p.price-discount)*pc.bought_qty
+            if pc.bought_qty > p.stock_remaining:
+                return {'message':'Some Products are out of Stock!'}
+            total_price+=(p.price)*pc.bought_qty 
             # add product to transaction
-            tp = TransactionProduct(t_id=id, p_id=p.p_id, bought_qty=pc.bought_qty,
-                                     paid=(p.price-discount)*pc.bought_qty, is_discount=is_discount )
+            tp = TransactionProduct(t_id=id, p_id=p.p_id, bought_qty=pc.bought_qty, paid=total_price)
             p.stock_remaining-=pc.bought_qty
             db.session.add(tp)
             db.session.delete(pc) #delete product from mycart
         # apply offer on total price if any
         cust_offer = CustomerOffers.query.filter_by(user_id=user_id).first()
+        offer_details={}
         if cust_offer is not None:
             offer = Offers.query.get(cust_offer.o_id)
             total_price-=offer.apply_discount(total_price)
+            offer_details={
+                'o_name': offer.o_name,
+                'discount': offer.discount,
+                'total_price':total_price,
+                'use_count_remaining': cust_offer.use_count-1
+            }
             cust_offer.use_count-=1
+            if (cust_offer.use_count==0):
+                # CALL CELERY TASK INTEREUPT TO DELETE OFFER FROM CUSTOMER_OFFER
+                pass
         date = datetime.now()
         t = Transaction(t_id=id, total_price=total_price, user_id=user_id, bought_date=date)
         db.session.add(t)
         db.session.commit() 
-        return {'message':'Order Placed!'}, 200
+        return {'message':'Order Placed!','offer':offer_details}, 200
